IdentityKit/IdentityKit/IdentityLibrary.cs
/// <summary>
/// کتابخانه مدیریت هویت برای سیستم شناسایی
/// </summary>
public class IdentityLibrary
{
    /// <summary>
    /// تابع ثبت‌نام کاربر جدید
    /// </summary>
    /// <param name="userName">نام کاربری کاربر</param>
    /// <param name="password">رمز عبور کاربر</param>
    /// <returns>نتیجه ثبت‌نام</returns>
    public RegistrationResult RegisterUser(string userName, string password)
    {
        // کد ثبت‌نام کاربر
        // ...

        return new RegistrationResult();
    }

    /// <summary>
    /// تابع ورود به سیستم
    /// </summary>
    /// <param name="userName">نام کاربری</param>
    /// <param name="password">رمز عبور</param>
    /// <returns>نتیجه ورود به سیستم</returns>
    public LoginResult Login(string userName, string password)
    {
        // کد ورود به سیستم
        // ...

        return new LoginResult();
    }

    // دیگر توابع و متدهای مد نظر

    /// <summary>
    /// تابع بازیابی رمز عبور با ایمیل
    /// </summary>
    /// <param name="email">آدرس ایمیل</param>
    /// <returns>نتیجه بازیابی رمز عبور</returns>
    public PasswordRecoveryResult RecoverPasswordByEmail(string email)
    {
        // کد بازیابی رمز عبور با ایمیل
        // ...

        return new PasswordRecoveryResult();
    }
}
