IdentityKit/IdentityKit/IdentityLibraryTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;

[TestClass]
public class IdentityLibraryTests
{
    [TestMethod]
    public void TestRegisterUser()
    {
        // Arrange
        var identityLibrary = new IdentityLibrary();
        var userName = "testUser";
        var password = "testPassword";

        // Act
        var result = identityLibrary.RegisterUser(userName, password);

        // Assert
        Assert.IsNotNull(result);
        // افزودن تست‌های دیگر بر اساس خروجی مورد انتظار
    }

    [TestMethod]
    public void TestLogin()
    {
        // Arrange
        var identityLibrary = new IdentityLibrary();
        var userName = "testUser";
        var password = "testPassword";

        // Act
        var result = identityLibrary.Login(userName, password);

        // Assert
        Assert.IsNotNull(result);
        // افزودن تست‌های دیگر بر اساس خروجی مورد انتظار
    }

    [TestMethod]
    public void TestRecoverPasswordByEmail()
    {
        // Arrange
        var identityLibrary = new IdentityLibrary();
        var email = "test@example.com";

        // Act
        var result = identityLibrary.RecoverPasswordByEmail(email);

        // Assert
        Assert.IsNotNull(result);
        // افزودن تست‌های دیگر بر اساس خروجی مورد انتظار
    }
}
